refactor + cluster awareness
